arm64: DT: msm8996-pm: Disable LPM prediction and do C4 retention
Disable the LPM prediction as, in reality, MSM8996 has only two
really usable PSCI C-States.... with the retention state that
we were previously using being really useless for both heat and
(obviously) power consumption.

Use the C4 state via hypervisor calls for retention (as it's not
getting in full C4 like this) which is as fast as going C2, but
saves the same amount of power.
The advantage of this is that when we are in retention and we
want to switch to full power collapse, we'll lose way less time.

Edit: Disable prediction via tunable.

Signed-off-by: xNombre <kartapolska@gmail.com>